requirements.txt
Mezzanine==3.1.9

drum

pillow

psycopg2

South

djangorestframework